#24 Create project structure for unit tests
Refactor project's structure
closes #24
