+++ pcbdevice/main.py
@@ -1,0 +1,6 @@
+from pcbdevice.utils.pbmformator import formatPbm
+from pcbdevice.utils.savetofile import matrixToFile
+
+if __name__ == "__main__":
+	# Usage example
+	matrixToFile(formatPbm('resources/original/test1ascii.pbm'), 'resources/reformatted/test1.csv')
