+++ pcbdevice/tests/utils/test_formatPbm.py
@@ -1,0 +1,5 @@
+from unittest import TestCase
+
+
+class TestFormatPbm(TestCase):
+	pass
